third_party/blink/web_tests/external/wpt/cookie-store/cookieStore_delete_arguments.https.any.serviceworker-expected.txt

This is a testharness.js-based test.
[FAIL] cookieStore.delete does not append / at the end of path
  assert_equals: expected null but got object "[object Object]"
Harness: the test ran to completion.
